// src/lib/api/generated/model/healthCheckerControllerCheck200.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { HealthCheckerControllerCheck200Details } from './healthCheckerControllerCheck200Details';
import type { HealthCheckerControllerCheck200Error } from './healthCheckerControllerCheck200Error';
import type { HealthCheckerControllerCheck200Info } from './healthCheckerControllerCheck200Info';

export type HealthCheckerControllerCheck200 = {
  status?: string;
  /** @nullable */
  info?: HealthCheckerControllerCheck200Info;
  /** @nullable */
  error?: HealthCheckerControllerCheck200Error;
  details?: HealthCheckerControllerCheck200Details;
};